+++ NEWS.md
@@ -14,6 +14,7 @@
 * Shorten waiting time in code explorer/package explorer by reading tarballs in memory (#707)
 * Move package upload process to a function
 * Updated logo in reports to lighter version to match application UI
+* Utilize `{archive}` to parse DESCRIPTION file in package upload process
 
 # riskassessment 3.0.0